+++ app/src/main/java/com/hansung/sherpa/MainActivity.kt
@@ -28,6 +28,7 @@ import com.hansung.sherpa.navigation.MyOnLocationChangeListener
 import com.hansung.sherpa.navigation.Navigation
 import com.hansung.sherpa.navigation.OnLocationChangeManager
 import com.hansung.sherpa.routelist.RouteListActivity
+import com.hansung.sherpa.searchscreen.SearchScreen
 import com.hansung.sherpa.ui.theme.SherpaTheme
 import com.naver.maps.geometry.LatLng
 import com.naver.maps.map.LocationTrackingMode
